drt: extract DrtModeRoutingModule
